+++ test/runtests.jl
@@ -49,6 +49,10 @@ TEST_CASES_DIR = pkgdir(XCALibre, "test/0_TEST_CASES")
         include("unit_test_wall_distance.jl")
     end
 
+    @testset "Wall Production Density Unit Test" begin
+        include("unit_test_wall_production_density.jl")
+    end
+
     # @testset "setFields Function Unit Test" begin
     #     include("unit_test_setFields.jl")
     # end